# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create": {
            "post": {
                "description": "It will create a product from provided information",
                "consumes": [
                    "application/json"
                ],
                "summary": "Create a product",
                "operationId": "create-product",
                "parameters": [
                    {
                        "description": "DisplayId of a product",
                        "name": "displayId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Name of a product",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Description of a product",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "250": {
                        "description": "A product is created. ErrorTypes array will be empty.",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errorTypes": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "251": {
                        "description": "All ready a product with displayId or name. ErrorType{1 =\u003e displayId, 2 =\u003e name}",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errorTypes": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/readdisplaycard": {
            "get": {
                "description": "It will collect primary data for display card for a product",
                "consumes": [
                    "application/json"
                ],
                "summary": "Read data for display card for a product",
                "operationId": "read-display-card-product",
                "parameters": [
                    {
                        "description": "Id of a product",
                        "name": "Id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "250": {
                        "description": "Display card data of a product",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "product": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "251": {
                        "description": "No product found with this id",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errorTypes": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    }
}